+++ src/client/components/app-footer/footer.ts
@@ -1,8 +1,8 @@
 import "./footer.scss";
 import template from "./footer.html";
-import { ComponentBase } from "@common/components";
 import badge from "../../assets/badge.png";
 import { singleton } from "tsyringe";
+import { ComponentBase } from "../components";
 
 @singleton()
 export class AppFooter extends ComponentBase {
