+++ src/client/components/page-chars/character-card/character_card.html
