+++ src/client/components/page-pols/polling/polling.scss
